move jabs-convert-parquet command to jabs-cli subcommand; update to use parquet output format from Alex's lixit detector
