+++ src/jabs/scripts/cli/cli.py
@@ -17,6 +17,7 @@ from jabs.classifier import Classifier
 from jabs.core.enums import ClassifierType, CrossValidationGroupingStrategy
 from jabs.project import Project, export_training_data, get_videos_to_prune
 
+from .convert_parquet import convert_parquet_command
 from .convert_to_nwb import run_conversion
 from .cross_validation import run_cross_validation
 from .postprocessing import apply_postprocessing_command
@@ -43,6 +44,7 @@ def cli(ctx: click.Context, verbose):
 
 
 cli.add_command(apply_postprocessing_command)
+cli.add_command(convert_parquet_command)
 cli.add_command(update_pose_command)
 cli.add_command(sample_pose_intervals_command)
 cli.add_command(sample_frames_command)
